+++ server/app.py
@@ -1,27 +1,16 @@
 from flask import Flask
-from flask_cors import CORS
-from flask_migrate import Migrate
-from flask_restful import Api
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import MetaData
-from flask_bcrypt import Bcrypt
-from config import get_config
+from flask_migrate import Migrate
+from config import Config
+
+# Import your models
+from server.models import db, User, Course, Lesson, Discussion, Enrollment
 
 app = Flask(__name__)
-app.config.from_object(get_config())
-app.json.compact = False
+app.config.from_object(Config)
 
-metadata = MetaData(naming_convention={
-    "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-})
-
-db = SQLAlchemy(metadata=metadata)
 db.init_app(app)
-
 migrate = Migrate(app, db)
-api = Api(app)
-CORS(app)
-bcrypt = Bcrypt(app)
 
 if __name__ == '__main__':
-    app.run(debug=app.config['DEBUG'])
+    app.run(debug=True)
