--- server/config.py
@@ -1,30 +1,0 @@
-import os
-
-class Config:
-    SECRET_KEY = os.environ.get('SECRET_KEY') or 'a_default_secret_key'
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or 'sqlite:///app.db'
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    CORS_HEADERS = 'Content-Type'
-    JSONIFY_PRETTYPRINT_REGULAR = False
-    DEBUG = False
-
-class DevelopmentConfig(Config):
-    DEBUG = True
-
-class TestingConfig(Config):
-    TESTING = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///test.db'
-
-class ProductionConfig(Config):
-    DEBUG = False
-
-def get_config():
-    env = os.getenv('FLASK_ENV', 'development')
-    if env == 'development':
-        return DevelopmentConfig
-    elif env == 'testing':
-        return TestingConfig
-    elif env == 'production':
-        return ProductionConfig
-    else:
-        return Config
