--- server/models/discussion.py
@@ -1,19 +1,0 @@
-from . import db
-
-class Discussion(db.Model):
-    tablename = 'discussions'
-
-    id = db.Column(db.Integer, primary_key=True)
-    topic = db.Column(db.Text, nullable=False)
-    content = db.Column(db.Text, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    course_id = db.Column(db.Integer, db.ForeignKey('courses.id'), nullable=False)
-    created_at = db.Column(db.DateTime, default=db.func.current_timestamp())
-    updated_at = db.Column(db.DateTime, default=db.func.current_timestamp(), onupdate=db.func.current_timestamp())
-
-    # Relationships
-    user = db.relationship('User', back_populates='discussions')
-    course = db.relationship('Course', back_populates='discussions')
-
-    def repr(self):
-        return f'<Discussion {self.topic}>'
